refactor: 调整类型定义的位置
